Bug 760956 - Part 1: Introduce variadic concatenateSelectionArgs r=rnewman
This allows us to more easily append multiple sets of args.

TODO: is it worth naming this appendSelectionArgs? We should keep the
2-param version appendSelectionArgs since it's more efficient for that
case (and the vast majority of uses are with 2-params) - it's probably
simpler for development to have both versions have the same name, and
as far as I can tell the compiler will prefer the non varargs version
when possible.

MozReview-Commit-ID: LcI6zNwpIxh

UltraBlame original commit: f0d9ab6ffd64bd61e90dc2a13120df985c91c823