Bug 1251870 - Disable indexing of objdir on Windows; r=ted
The Windows content indexing service has been known to scan the objdir.
This can add significant system overhead and slow down builds or
subsequent operations. The objdir is meant to be a short-lived black box
and there really isn't a major benefit to indexing it.

There is a file attribute on Windows that disables content indexing.
This commit adds a utility function for creating a directory and
optionally disabling indexing on it. We call it at the top of
`mach build` to ensure the objdir as content indexing disabled.

MozReview-Commit-ID: 68gxCxbkVAN

UltraBlame original commit: b7769b2dc2bd9d06f67c5f8c1121d3d29d5acc6a