Bug 1252901 - don't use ScopedFreePtrTraits in MediaEngineWrapper.h; r=jesup
ScopedFreePtrTraits is somewhat of an internal implementation detail of
Scoped.h, and it's not hard to write out the two static functions we
need from ScopedFreePtrTraits anyway.  Removing this means that we can
clear the way for ScopedFreePtr to be removed.

UltraBlame original commit: 29dec7e96a9c87e54b5ec1b169df10c94c299f53