Bug 946857 - Part 2: Add Android LoginsProvider. r=nalexander
LoginsProvider is an all-Android implementation of PasswordsProvider.
PasswordsProvider is an SQLite database backed by the version of
SQLite that ships with Gecko.  It is concurrently accessed from Gecko
and it runs with a special lifecycle that includes a separate
heavy-weight process.  Eventually we'll migrate the Gecko-side
passwords interface to use the new Android-side LoginsProvider, but
for now we just want to get the new provider landed and the tests
running.

MozReview-Commit-ID: Bx19D68tMtI
***
Bug 946857 - Fold into part2: review nits.

MozReview-Commit-ID: LmPwIvebfrr

UltraBlame original commit: 2c6222ddc0330ef5b0473363889a46110a8ade07