Bug 1256028 - Disable C4477 to unblock compilation on VS2015; r=mcmanus
As part of unblocking building with VS2015u1 in automation, I'm mass
disabling compiler warnings that are turned into errors. This is not
the preferred mechanism to fix compilation warnings. So hopefully
this patch never lands because someone insists of fixing the underlying
problem instead. But if it does land, hopefully the workaround is
only temporary.

MozReview-Commit-ID: 9AkkAUDMln6

UltraBlame original commit: cdea1598a698b6ba2024ad8c77a093800a6d80ae