Bug 1258452 - Explicitly request READ_EXTERNAL_STORAGE permission. r=sebastian
This permission was added in API 16, and is only enforced in API 19+.
(It's benign to add it to APKs installed on API 15.)

We want to declare it explicitly so that the bouncer APK and the main
APK have the same permission set.  There appears to be some fanciness
with Gradle's implied permission system where the bouncer APK does not
request READ implicitly where-as the main APK does request READ
implicitly.  This just makes things explicit (and uniform).

MozReview-Commit-ID: LBblAPa0duV

UltraBlame original commit: a42ef90399bd6a848fb839647ea328e9453ffe5f