Bug 1258615 - Remove the "magic" that sets a result from set_config() in depends functions. r=chmanchester
Currently, if a depends function doesn't have a return statement or
return None, a result is automatically set from all the set_config()
called from the function.

As we're going to move set_config to the global namespace, and as this
feature is only used once, and it's only used for something that was
written before ReadOnlyNamespace was exposed to the sandbox, we can
"safely" get rid of it.

UltraBlame original commit: 1f6d8c58450fd07f9962e1d3b5b80d74096e4755