Bug 1258769 - Use soter to add checkstyle gradle task with simple checks. r=nalexander
Run checkstyle via:
 ./mach gradle app:checkstyle

This is intended to be added to automation in bug 1258787.

Concerns with this patch:
  1) I don't have a maven-metadata-local.xml. However, I didn't take the
snapshotted version and [1] seems to indicate it's unnecessary for my
particular build.

For the addition of future checks, it's worth noting Google's config is
available:
  https://github.com/checkstyle/checkstyle/blob/3e4367941c3e9680703e8ea8400abbd5dc78e1d9/src/main/resources/google_checks.xml

And this version contains links with descriptions of each of the tasks:
  http://checkstyle.sourceforge.net/google_style.html

[1]: https://maven.apache.org/ref/3.3.3/maven-repository-metadata/

MozReview-Commit-ID: ID3X9ZA27b0

UltraBlame original commit: fe11700b6cdfdb0fca3a9b2a54bf6d339792df24