Bug 1234622. Tweak how nsDocumentViewer::FindContainerView finds the parent presshell. r=bz
During page transition the presshell we get via the parent docshell is the presshell for the _next_ document.

Whereas, the presshell we get via the containing frame element (iframe etc) is the current document.

If we use the presshell for the next document to get the primary frame for the containerElement it will return null because it doesn't have the right document.

UltraBlame original commit: ff20990e88e15a5002d63b2cfd8aa390138c1cb4