Bug 1260261 - Associate a `CensusTreeNode` with the leaf in the census report from which it was generated; r=jimb
This commit gives every `CensusTreeNode` instance a `reportLeafIndex` member
that is an index into a pre-order traversal of the census report from which it
was generated. This can be used to get the leaf in the census report
corresponding to a given `CensusTreeNode` instance.

UltraBlame original commit: f3ea2fdec3196f6a498c7fe8c1e48e460331f602