Bug 1256573 - Add a new imports primitive that allows to import modules into the decorated functions. r=nalexander
Currently, we have advanced, that gives the decorated functions access
to all the builtins and consequently, to the import statement.
That is a quite broad approach and doesn't allow to easily introspect
what functions are importing which modules.

This change introduces a new decorator that allows to import modules one
by one into the decorated functions.

Note: by the end of the change series, advanced will be gone.

UltraBlame original commit: 96ef4c62378d5ddabb8c27c745bc928517273c20